# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Registry Service",
    "description": "Aggregates OpenAPI specs from multiple microservices into a single queryable registry. Provides REST endpoints and MCP tools for LLM-powered service discovery.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "schemas": {
...
      "ServicesResponse": {
        "type": "object",
        "properties": {
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceEntry"
            }
          }
        },
        "required": [
          "services"
        ]
      },
...
      "UnauthorizedError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Invalid or missing API key"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "OpenApiSpec": {
        "type": "object",
        "properties": {},
        "additionalProperties": {
          "nullable": true
        },
        "description": "An OpenAPI specification object"
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "available": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "error"
        ]
      },
      "UpstreamError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "detail"
        ]
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string"
          },
          "resultCount": {
            "type": "number"
          },
          "indexSize": {
            "type": "number"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            }
          }
        },
        "required": [
          "query",
          "resultCount",
          "indexSize",
          "results"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/services": {
      "get": {
        "summary": "List all registered services",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service (optional for read-only endpoints)"
            },
            "required": false,
            "description": "Internal org UUID from client-service (optional for read-only endpoints)",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service (optional for read-only endpoints)"
            },
            "required": false,
            "description": "Internal user UUID from client-service (optional for read-only endpoints)",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "List of registered services",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
    "/openapi/{service}": {
      "get": {
        "summary": "Get OpenAPI spec for a specific service",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Service name"
            },
            "required": true,
            "description": "Service name",
            "name": "service",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service (optional for read-only endpoints)"
            },
            "required": false,
            "description": "Internal org UUID from client-service (optional for read-only endpoints)",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service (optional for read-only endpoints)"
            },
            "required": false,
            "description": "Internal user UUID from client-service (optional for read-only endpoints)",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OpenAPI spec for the service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenApiSpec"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Service not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Failed to fetch spec from upstream",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpstreamError"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Search for API endpoints across all services using ranked full-text search. Supports fuzzy matching and prefix search.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Search query (e.g. 'send email', 'brand extract')"
            },
            "required": true,
            "description": "Search query (e.g. 'send email', 'brand extract')",
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter to a specific service"
            },
            "required": false,
            "description": "Filter to a specific service",
            "name": "service",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by HTTP method (e.g. 'POST')"
            },
            "required": false,
            "description": "Filter by HTTP method (e.g. 'POST')",
            "name": "method",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by path prefix (e.g. '/v1/')"
            },
            "required": false,
            "description": "Filter by path prefix (e.g. '/v1/')",
            "name": "pathPrefix",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Max results (omit for all results)"
            },
            "required": false,
            "description": "Max results (omit for all results)",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service (optional for read-only endpoints)"
            },
            "required": false,
            "description": "Internal org UUID from client-service (optional for read-only endpoints)",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service (optional for read-only endpoints)"
            },
            "required": false,
            "description": "Internal user UUID from client-service (optional for read-only endpoints)",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
    "/llm-context": {
      "get": {
        "summary": "LLM-friendly overview: list all services with name, description, and endpoint count. Use GET /llm-context/{service} for endpoint details.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service (optional for read-only endpoints)"
            },
            "required": false,
            "description": "Internal org UUID from client-service (optional for read-only endpoints)",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service (optional for read-only endpoints)"
            },
            "required": false,
            "description": "Internal user UUID from client-service (optional for read-only endpoints)",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Lightweight service overview for LLM consumption",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmOverviewResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
    "/llm-context/{service}": {
      "get": {
        "summary": "LLM-friendly endpoint list for a specific service. Supports filtering by method, path group, or path prefix. Large services (30+ endpoints) auto-group by path prefix.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Service name"
            },
            "required": true,
            "description": "Service name",
            "name": "service",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by HTTP method (e.g. 'POST')"
            },
            "required": false,
            "description": "Filter by HTTP method (e.g. 'POST')",
            "name": "method",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by path group (e.g. 'campaigns', 'brands')"
            },
            "required": false,
            "description": "Filter by path group (e.g. 'campaigns', 'brands')",
            "name": "group",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by path prefix (e.g. '/v1/campaigns')"
            },
            "required": false,
            "description": "Filter by path prefix (e.g. '/v1/campaigns')",
            "name": "pathPrefix",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service (optional for read-only endpoints)"
            },
            "required": false,
            "description": "Internal org UUID from client-service (optional for read-only endpoints)",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service (optional for read-only endpoints)"
            },
            "required": false,
            "description": "Internal user UUID from client-service (optional for read-only endpoints)",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Endpoint list for the service (flat or grouped depending on size)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmServiceDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Service not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Failed to fetch spec from upstream",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "detail"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/openapi.json": {
      "get": {
        "summary": "This service's own OpenAPI specification",
        "responses": {
          "200": {
            "description": "OpenAPI 3.0 specification",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenApiSpec"
                }
              }
            }
          },
          "404": {
            "description": "Spec not generated yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/discover/services": {
      "get": {
        "summary": "Level 1: every running service with a one-line description",
        "description": "Start here to explore the platform. Each service: name, one-line description, endpoint count. Next: GET /discover/services/{service}/endpoints.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Text filter on name + description; every word must match"
            },
            "required": false,
            "description": "Text filter on name + description; every word must match",
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Max services returned (default all)"
            },
            "required": false,
            "description": "Max services returned (default all)",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Services",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverServicesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}